requirements.txt
django=4.2
python=3.10
channels=3.0.4
channels-redis=3.3.1
requests
# networkx~=2.4
numpy
selenium=3.141.0
django-polymorphic
django-tinymce4-lite
pillow
django-smart-selects